split proto library for future deps usage (#4239)
