+++ ydb/core/formats/arrow/protos/ya.make
@@ -1,0 +1,11 @@
+PROTO_LIBRARY()
+
+SRCS(
+    ssa.proto
+)
+
+PEERDIR(
+
+)
+
+END()
