+++ ydb/core/formats/arrow/ssa_program_optimizer.h
@@ -2,7 +2,6 @@
 
 #include "program.h"
 
-#include <ydb/core/protos/ssa.pb.h>
 #include <ydb/core/tablet_flat/flat_dbase_scheme.h>
 
 namespace NKikimr::NSsa {
