+++ ydb/core/kqp/compute_actor/ya.make
@@ -21,6 +21,7 @@ PEERDIR(
     ydb/core/kqp/runtime
     ydb/core/tx/datashard
     ydb/core/tx/scheme_cache
+    ydb/core/formats/arrow/protos
     ydb/library/yql/dq/actors/compute
     ydb/library/yql/providers/generic/actors
     ydb/library/yql/providers/s3/actors
