+++ ydb/core/kqp/query_compiler/kqp_olap_compiler.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <ydb/core/kqp/expr_nodes/kqp_expr_nodes.h>
-#include <ydb/core/protos/ssa.pb.h>
 
 #include <ydb/core/kqp/provider/yql_kikimr_gateway.h>
