+++ ydb/core/protos/tx_columnshard.proto
@@ -1,9 +1,10 @@
 import "ydb/library/actors/protos/actors.proto";
 import "ydb/core/tx/columnshard/engines/scheme/statistics/protos/data.proto";
+import "ydb/core/tx/columnshard/engines/protos/portion_info.proto";
 import "ydb/core/protos/flat_scheme_op.proto";
 import "ydb/core/protos/long_tx_service.proto";
 import "ydb/core/protos/subdomains.proto";
-import "ydb/core/protos/ssa.proto";
+import "ydb/core/formats/arrow/protos/ssa.proto";
 
 package NKikimrTxColumnShard;
 option java_package = "ru.yandex.kikimr.proto";
@@ -269,33 +270,6 @@ message TSchemaSeqNo {
 
 message TIndexGranuleMeta {
     optional uint32 MarkSize = 1; // Composite key mark (granule border) size: count of first PK elements in mark
-}
-
-message TSnapshot {
-    optional uint64 PlanStep = 1;
-    optional uint64 TxId = 2;
-}
-
-message TIndexPortionMeta {
-    oneof Produced {
-        bool IsInserted = 1;
-        bool IsCompacted = 2;
-        bool IsSplitCompacted = 3;
-        bool IsEvicted = 4;
-    }
-    optional string TierName = 5;
-    optional bytes PrimaryKeyBorders = 6; // arrow::RecordBatch with first and last ReplaceKey rows
-    optional TSnapshot RecordSnapshotMin = 7;
-    optional TSnapshot RecordSnapshotMax = 8;
-    optional NKikimrColumnShardStatisticsProto.TPortionStorage StatisticsStorage = 9;
-}
-
-message TIndexColumnMeta {
-    optional uint32 NumRows = 1;
-    optional uint32 RawBytes = 2;
-    optional NKikimrSSA.TProgram.TConstant MinValue = 3;
-    optional NKikimrSSA.TProgram.TConstant MaxValue = 4;
-    optional TIndexPortionMeta PortionMeta = 5; // First PK column could contain portion info
 }
 
 message TSchemaTxBody {
