+++ ydb/core/protos/ya.make
@@ -116,7 +116,6 @@ SRCS(
     serverless_proxy_config.proto
     shared_cache.proto
     sqs.proto
-    ssa.proto
     statestorage.proto
     statistics.proto
     stream.proto
@@ -167,6 +166,8 @@ PEERDIR(
     ydb/library/services
     ydb/library/ydb_issue/proto
     ydb/core/tx/columnshard/engines/scheme/statistics/protos
+    ydb/core/tx/columnshard/engines/protos
+    ydb/core/formats/arrow/protos
 )
 
 CPP_PROTO_PLUGIN0(config_proto_plugin ydb/core/config/tools/protobuf_plugin)
