+++ ydb/core/tx/columnshard/common/scalars.cpp
@@ -1,7 +1,6 @@
 #include "scalars.h"
 
 #include <ydb/core/formats/arrow/switch_type.h>
-#include <ydb/core/protos/ssa.pb.h>
 #include <ydb/library/yverify_stream/yverify_stream.h>
 #include <util/system/unaligned_mem.h>
