+++ ydb/core/tx/columnshard/common/tests/ya.make
@@ -5,7 +5,7 @@ SRCS(
 )
 
 PEERDIR(
-    ydb/core/protos
+    ydb/core/formats/arrow/protos
     contrib/libs/apache/arrow
     ydb/core/formats/arrow
     ydb/core/kqp/compute_actor
