+++ ydb/core/tx/columnshard/common/ya.make
@@ -11,7 +11,7 @@ SRCS(
 )
 
 PEERDIR(
-    ydb/core/protos
+    ydb/core/formats/arrow/protos
     contrib/libs/apache/arrow
     ydb/core/formats/arrow
     ydb/core/tx/columnshard/common/protos
