+++ ydb/core/tx/columnshard/data_sharing/protos/sessions.proto
@@ -57,3 +57,7 @@ message TSourceSession {
         repeated TPathPortionsHash PathHashes = 7;
     }
 }
+
+message TEvProposeFromInitiator {
+    optional TDestinationSession Session = 1;
+}
