+++ ydb/core/tx/columnshard/data_sharing/protos/ya.make
@@ -9,10 +9,10 @@ SRCS(
 )
 
 PEERDIR(
+    ydb/core/tx/columnshard/engines/protos
     ydb/core/tx/columnshard/common/protos
     ydb/library/actors/protos
     ydb/core/tx/columnshard/blobs_action/protos
-    ydb/core/protos
 
 )
