+++ ydb/core/tx/columnshard/engines/portions/ya.make
@@ -18,6 +18,7 @@ PEERDIR(
     ydb/core/tx/columnshard/splitter
     ydb/core/tx/columnshard/common
     ydb/core/tx/columnshard/data_sharing/protos
+    ydb/core/tx/columnshard/engines/protos
     ydb/core/tablet_flat
 )
