+++ ydb/core/tx/columnshard/engines/protos/ya.make
@@ -1,0 +1,13 @@
+PROTO_LIBRARY()
+
+SRCS(
+    portion_info.proto
+)
+
+PEERDIR(
+    ydb/core/tx/columnshard/engines/scheme/statistics/protos
+    ydb/core/formats/arrow/protos
+
+)
+
+END()
