+++ ydb/core/tx/program/ya.make
@@ -8,6 +8,7 @@ SRCS(
 PEERDIR(
     ydb/core/formats/arrow
     ydb/core/protos
+    ydb/core/formats/arrow/protos
     ydb/core/tablet_flat
     ydb/library/yql/minikql/comp_nodes
     ydb/library/yql/core/arrow_kernels/registry
